Issue#1228: Reworked rule numbers to reflect execution phases.
