+++ tests/subtype/rule_005_test_input.fixed.vhd
@@ -5,7 +5,13 @@ architecture RTL of FIFO is
 
   -- Violations below
 
-  subtype counter is unsigned(4 downto 0);
+  subtype counter
+ is unsigned(4 downto 0);
+
+  subtype counter
+
+
+  is unsigned(4 downto 0);
 
 begin
