--- tests/subtype/rule_013_test_input.fixed_upper.vhd
@@ -1,12 +1,0 @@
-
-architecture RTL of FIFO is
-
-  SUBTYPE state_machine is (idle, write, read, done);
-
-  -- Violations below
-
-  SUBTYPE state_machine is (idle, write, read, done);
-
-begin
-
-end architecture RTL;
