--- tests/subtype/rule_013_test_input.vhd
@@ -1,12 +1,0 @@
-
-architecture RTL of FIFO is
-
-  subtype state_machine is (idle, write, read, done);
-
-  -- Violations below
-
-  SUBTYPE state_machine is (idle, write, read, done);
-
-begin
-
-end architecture RTL;
