+++ tests/subtype/rule_202_test_input.vhd
