+++ tests/subtype/test_rule_005.py
@@ -11,15 +11,18 @@ sTestDir = os.path.dirname(__file__)
 
 lFile, eError = vhdlFile.utils.read_vhdlfile(os.path.join(sTestDir, "rule_005_test_input.vhd"))
 
+dIndentMap = utils.read_indent_file()
+
 lExpected = []
 lExpected.append("")
 utils.read_file(os.path.join(sTestDir, "rule_005_test_input.fixed.vhd"), lExpected)
 
 
-class test_subtype_definition_rule(unittest.TestCase):
+class test_type_definition_rule(unittest.TestCase):
     def setUp(self):
         self.oFile = vhdlFile.vhdlFile(lFile)
         self.assertIsNone(eError)
+        self.oFile.set_indent_map(dIndentMap)
 
     def test_rule_005(self):
         oRule = subtype.rule_005()
@@ -27,7 +30,7 @@ class test_subtype_definition_rule(unittest.TestCase):
         self.assertEqual(oRule.name, "subtype")
         self.assertEqual(oRule.identifier, "005")
 
-        lExpected = [8]
+        lExpected = [8, 11]
 
         oRule.analyze(self.oFile)
         self.assertEqual(lExpected, utils.extract_violation_lines_from_violation_object(oRule.violations))
