+++ tests/subtype/test_rule_202.py
@@ -9,11 +9,11 @@ from vsg.rules import subtype
 
 sTestDir = os.path.dirname(__file__)
 
-lFile, eError = vhdlFile.utils.read_vhdlfile(os.path.join(sTestDir, "rule_008_test_input.vhd"))
+lFile, eError = vhdlFile.utils.read_vhdlfile(os.path.join(sTestDir, "rule_202_test_input.vhd"))
 
 lExpected = []
 lExpected.append("")
-utils.read_file(os.path.join(sTestDir, "rule_008_test_input.fixed.vhd"), lExpected)
+utils.read_file(os.path.join(sTestDir, "rule_202_test_input.fixed.vhd"), lExpected)
 
 
 class test_type_definition_rule(unittest.TestCase):
@@ -21,19 +21,19 @@ class test_type_definition_rule(unittest.TestCase):
         self.oFile = vhdlFile.vhdlFile(lFile)
         self.assertIsNone(eError)
 
-    def test_rule_008(self):
-        oRule = subtype.rule_008()
+    def test_rule_202(self):
+        oRule = subtype.rule_202()
         self.assertTrue(oRule)
         self.assertEqual(oRule.name, "subtype")
-        self.assertEqual(oRule.identifier, "008")
+        self.assertEqual(oRule.identifier, "202")
 
         lExpected = [10]
 
         oRule.analyze(self.oFile)
         self.assertEqual(lExpected, utils.extract_violation_lines_from_violation_object(oRule.violations))
 
-    def test_fix_rule_008(self):
-        oRule = subtype.rule_008()
+    def test_fix_rule_202(self):
+        oRule = subtype.rule_202()
 
         oRule.fix(self.oFile)
