+++ vsg/rules/subtype/__init__.py
@@ -6,10 +6,10 @@ from .rule_003 import rule_003
 from .rule_004 import rule_004
 from .rule_005 import rule_005
 from .rule_006 import rule_006
-from .rule_007 import rule_007
-from .rule_008 import rule_008
-from .rule_009 import rule_009
-from .rule_010 import rule_010
+from .rule_101 import rule_101
+from .rule_102 import rule_102
+from .rule_201 import rule_201
+from .rule_202 import rule_202
 from .rule_100 import rule_100
 from .rule_200 import rule_200
 from .rule_500 import rule_500
