+++ vsg/rules/subtype/rule_005.py
@@ -1,30 +1,29 @@
 # -*- coding: utf-8 -*-
 
-from vsg import token
-from vsg.rules.whitespace_between_token_pairs import Rule
 
-lTokens = []
-lTokens.append([token.subtype_declaration.identifier, token.subtype_declaration.is_keyword])
+from vsg.rules import move_token_next_to_another_token as Rule
+from vsg.token import subtype_declaration as token
 
 
 class rule_005(Rule):
     """
-    This rule checks for a single space before the **is** keyword.
-
-    |configuring_whitespace_rules_link|
+    This rule checks the identifier is on the same line as the **subtype** keyword.
 
     **Violation**
 
     .. code-block:: vhdl
 
-       subtype counter    is unsigned(4 downto 0);
+       subtype
+       st_counter is
 
     **Fix**
 
     .. code-block:: vhdl
 
-       subtype counter is unsigned(4 downto 0);
+       subtype st_counter
+       is
     """
 
     def __init__(self):
-        super().__init__(lTokens)
+        super().__init__(token.subtype_keyword, token.identifier)
+        self.solution = "Ensure identifier is on the same line as the *subtype* keyword."
